+++ forward_engineering/generateContainerScript.js
@@ -1,0 +1,68 @@
+const { get } = require('lodash');
+const logHelper = require('../shared/helpers/logHelper');
+const ForwardEngineeringScriptBuilder = require('./services/forwardEngineeringScriptBuilder');
+const { GENERATING_CONTAINER_SCRIPT } = require('../shared/enums/staticMessages');
+const { includeSamples } = require('./utils/includeSamples');
+
+/**
+ * @param {ConnectionInfo} connectionInfo
+ * @param {AppLogger} appLogger
+ * @param {Callback} callback
+ * @param {App} app
+ */
+const generateContainerScript = async (connectionInfo, appLogger, callback, app) => {
+	const logger = logHelper.createLogger({
+		title: GENERATING_CONTAINER_SCRIPT,
+		hiddenKeys: connectionInfo.hiddenKeys,
+		logger: appLogger,
+	});
+
+	try {
+		const scriptBuilder = new ForwardEngineeringScriptBuilder();
+
+		const { jsonData, collections, options } = connectionInfo;
+		const { origin, additionalOptions } = options;
+		const rawScope = get(connectionInfo.containerData, '[0]', {});
+		const scope = {
+			...rawScope,
+			bucketName: rawScope?.bucket ?? '',
+		};
+		const collectionsData = collections.map(schema => ({
+			...JSON.parse(schema),
+			namespace: scope?.namespace,
+			bucketName: scope?.bucketName,
+			scopeName: scope?.name,
+		}));
+
+		scriptBuilder.addScopeScript(scope);
+		collectionsData.forEach(collection => scriptBuilder.addCollectionScripts(collection));
+
+		if (!includeSamples(additionalOptions)) {
+			const { script } = scriptBuilder.buildScriptSeparateFromInsertScripts();
+			return callback(null, script);
+		}
+
+		scriptBuilder.addContainerInsertScripts({ collections: collectionsData, jsonData });
+
+		if (origin !== 'ui') {
+			return callback(null, scriptBuilder.buildScriptConcatenatedWithInsertScripts('\n\n'));
+		}
+
+		const { script, insertScripts } = scriptBuilder.buildScriptSeparateFromInsertScripts();
+		callback(null, [
+			{ title: 'Couchbase script', script },
+			{
+				title: 'Sample data',
+				script: insertScripts,
+			},
+		]);
+	} catch (error) {
+		logger.error(error);
+
+		callback({ message: error.message, stack: error.stack });
+	}
+};
+
+module.exports = {
+	generateContainerScript,
+};
