+++ forward_engineering/utils/includeSamples.js
@@ -1,0 +1,6 @@
+const includeSamples = (additionalOptions = []) =>
+	Boolean(additionalOptions.find(option => option.id === 'INCLUDE_SAMPLES' && option.value));
+
+module.exports = {
+	includeSamples,
+};
